redo db models for search
